split output loading functions into raw vs. transformed
